chore: update forwarding API stubs
